requirements-pip.txt
anyio==4.6.2
Brotli==1.0.9
certifi==2025.1.31
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.7
iniconfig==1.1.1
lxml==5.3.0
OWSLib==0.24.1
packaging==24.2
pip==25.0
pluggy==1.5.0
pyproj==3.6.1
PySocks==1.7.1
pytest==8.3.4
python-box==7.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==75.8.0
six==1.16.0
sniffio==1.3.0
toml==0.10.2
urllib3==2.3.0
wheel==0.45.1
